Move navigation to a new file for code separating
